move get_redis_broker helper to utils and test
